Revise Model/Keyfile
First, we move `::jsConfig()` and `::jsL10n()` to more appropriate
places, since we don't want to blur the Model's responsibilites.

Then we see that there is no need for Keyfile; it was meant to separate
the actual file system access, but since we have vfsStream, we can use
it to test the Model directly, as soon as we inject the current time
into the object.

What's left is the vacuous name `Model`, but for now we can't come up
with something better (an obvious idea is `Keymaster`, but that would
require to rename some methods); we leave that for another day.

* Remove unused method
* Rewrite ModelTest against VFS
* Inject current time into Model
* Don't clear the whole statcache (no longer necessary as of PHP 5.3.0)
* Inline Keyfile
* Simplify Keyfile::purge() and ::extend()
* Move Model::jsL10n() to View::jsTexts()
* Move Model::jsConfig() to EmitScripts::jsConf()