Revise tests
When testing infrastructure classes, we test against a test-specific
subclass of the SUT (unless we can use `vsfStream` or such), which
doubles internal thin wrappers over untestable DOC APIs.  We mark these
internal thin wrappers as untestable.

When testing application layer classes, we must only double the public
API of infrastructure classes; otherwise we couple the tests to
implementation details, which is not desirable.

* Only stub public API of DOCs
* Test Request using a test-specific subclass of the SUT
* Only stub public API of DOCs (partly revert c8195ed)