Separate responsibilities of Response
As is, `Response` is actually a hybrid of a value object, with an
additional `::respond()` method.  The latter should be moved to an own
object, `Responder`, for a clean separation.

As can be seen, this allows us to test a test-specific subclass of
`Responder`, but that requires a shitload of boilerplate code, while
the test is still pretty meaningless (it's almost an exact duplication
of the implementation).

If we would go on with this, `Response` should be moved to the `Value`
subnamespace, but at least for now, we stick with what we had.